Sun, Feb 12, 2023, 2:20 PM +03:00

+++ calculator.c
@@ -8,12 +8,12 @@ int main (void)
    // float a = x+y;
    // float b = x*y;
     //float c = x-y;
-    float d = x/y;
+    float d = (float)x/(float)y;
     //int e = x%y;
 
     //printf("Sum of %.2f and %f is %f\n",x,y,a);
     //printf("Multiplication of %.2f and %f is %f\n",x,y,b);
     //printf("Substraction of %f and %.f is %f\n",x,y,c);
-    printf("Division of %f and %f is %.2f\n",x,y,d);
+    printf("Division of %i and %i is %.2f\n",x,y,d);
     //printf("Remainder of %f and %f is %f",x,y,e);
 }
